+++ packages/nodes-base/nodes/Misp/descriptions/index.ts
@@ -4,6 +4,7 @@ export * from './EventTagDescription';
 export * from './FeedDescription';
 export * from './GalaxyDescription';
 export * from './NoticelistDescription';
+export * from './ObjectDescription';
 export * from './OrganisationDescription';
 export * from './TagDescription';
 export * from './UserDescription';
